requirements.txt
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
Flask-Minify==0.32
Flask-PyMongo==2.3.0
Flask-Session==0.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
passlib==1.7.4
pymongo==3.12.0
Werkzeug==2.0.2
python-dotenv==0.19.1